add prettier config to generator add openapi.ts check for latest version
